Display handler cleanup
