+++ libs/display.py
@@ -1,5 +1,9 @@
 """Sets the values for displaying supported chars on a Waveshare RGB LED Pico hat (in landscape)"""
 
+import re
+
+from wsgi_web_app_helpers import get_json_wsgi_input, bad_request
+
 
 class TooManyCharsException(Exception):
     """Oh no, there are too many characters"""
@@ -64,3 +68,82 @@ def gen_char_values(chars):
 
     return active_pixels
 
+
+def clear(neo):
+    """Clears the screen"""
+    neo.fill([0, 0, 0])
+    neo.show()
+    return None, None
+
+
+class BaseDisplayHandler:
+    """Base class for handling display controls from JSON POST request data."""
+
+    def __init__(self, json_request, neo, num_px):
+        self.response = None
+        self.status_code = None
+        self.req_data = get_json_wsgi_input(json_request, bad_request)
+        self.neo = neo
+        self.num_px = num_px
+
+    def clear(self):
+        return clear(self.neo)
+
+    def set(self):
+        raise NotImplementedError("set method needs implementing.")
+
+
+class SetPx(BaseDisplayHandler):
+
+    def set(self):
+        if not isinstance(self.req_data, dict):  # Something bad has happened here.
+            return self.req_data
+
+        self.clear()
+        for rgb, pixels in self.req_data.items():
+            if re.match("\d\d\d", rgb) and isinstance(pixels, list):
+                for px in pixels:
+                    try:
+                        px_int = int(px)
+                    except Exception:
+                        return bad_request("Pixel numbers from list must be integers")
+                    if px_int > self.num_px:
+                        return bad_request(
+                            "Cannot set value to pixel  %s. Max supported: %s" % (px_int, self.num_px)
+                        )
+
+                    self.neo[int(px)] = [int(x * 200) for x in rgb]
+            else:
+                return bad_request('Input did not match schema. Example: {"010": [1,2,3,4,5], "011": [16,60]}')
+
+        self.neo.show()
+        return self.response, self.status_code
+
+
+class SetTemp(BaseDisplayHandler):
+
+    def set(self):
+        if not isinstance(self.req_data, dict):  # Something bad has happened here.
+            return self.req_data
+
+        for rgb, temp_value in self.req_data.items():
+            # make sure we are a string
+            temp_value = str(temp_value)
+            # simple regex for temp. "--" is a valid value for null, \d- is clearly just nonsense.
+            # regex's are a little hideous as MicroPython's regex engine isn't too glamorous...
+            if re.match("\d\d\d", rgb) and re.match("(--)|(^-\d)|(\d\d)|(\d)", temp_value):
+                try:
+                    rgb_temp_vals = gen_char_values(temp_value)
+                except Exception as exc:
+                    return bad_request(repr(exc))
+
+                self.clear() # only clearing the screen now as success seems likely
+                for px in rgb_temp_vals:
+                    self.neo[int(px)] = [int(x * 200) for x in rgb]
+            else:
+                return bad_request(
+                    "Input value %s doesn't appear to match the expected format: \{'001': '-3'\}." % self.req_data
+                )
+
+        self.neo.show()
+        return self.response, self.status_code
